+++ bin/check-mysql-replication-status.rb
@@ -24,7 +24,7 @@
 #
 
 require 'sensu-plugin/check/cli'
-require 'mysql2'
+require 'mysql'
 require 'inifile'
 
 class CheckMysqlReplicationStatus < Sensu::Plugin::Check::CLI
@@ -103,11 +103,11 @@ class CheckMysqlReplicationStatus < Sensu::Plugin::Check::CLI
     end
 
     begin
-      db = Mysql2::Client.new(host: db_host, username: db_user, password: db_pass, port: config[:port].to_i, socket: config[:socket])
+      db = Mysql.new(db_host, db_user, db_pass, nil, config[:port], config[:socket])
       results = db.query 'show slave status'
 
       unless results.nil?
-        results.each do |row|
+        results.each_hash do |row|
           warn "couldn't detect replication status" unless
             %w(Slave_IO_State Slave_IO_Running Slave_SQL_Running Last_IO_Error Last_SQL_Error Seconds_Behind_Master).all? do |key|
               row.key? key
@@ -141,7 +141,7 @@ class CheckMysqlReplicationStatus < Sensu::Plugin::Check::CLI
         ok 'show slave status was nil. This server is not a slave.'
       end
 
-    rescue Mysql2::Error => e
+    rescue Mysql::Error => e
       errstr = "Error code: #{e.errno} Error message: #{e.error}"
       critical "#{errstr} SQLSTATE: #{e.sqlstate}" if e.respond_to?('sqlstate')