+++ bin/metrics-mysql-graphite.rb
@@ -31,11 +31,11 @@
 #
 
 require 'sensu-plugin/metric/cli'
-require 'mysql2'
+require 'mysql'
 require 'socket'
 require 'inifile'
 
-class Mysql2Graphite < Sensu::Plugin::Metric::CLI::Graphite
+class MysqlGraphite < Sensu::Plugin::Metric::CLI::Graphite
   option :host,
          short: '-h HOST',
          long: '--host HOST',
@@ -205,20 +205,14 @@ class Mysql2Graphite < Sensu::Plugin::Metric::CLI::Graphite
         db_pass = config[:password]
       end
       begin
-        mysql = Mysql2::Client.new(
-          host: mysql_host,
-          port: config[:port],
-          username: db_user,
-          password: db_pass,
-          socket: config[:socket]
-        )
+        mysql = Mysql.new(mysql_host, db_user, db_pass, nil, config[:port], config[:socket])
 
         results = mysql.query('SHOW GLOBAL STATUS')
       rescue => e
         puts e.message
       end
 
-      results.each do |row|
+      results.each_hash do |row|
         metrics.each do |category, var_mapping|
           if var_mapping.key?(row['Variable_name'])
             output "#{config[:scheme]}.#{mysql_shorthostname}.#{category}.#{var_mapping[row['Variable_name']]}", row['Value']
@@ -230,7 +224,7 @@ class Mysql2Graphite < Sensu::Plugin::Metric::CLI::Graphite
         slave_results = mysql.query('SHOW SLAVE STATUS')
         # should return a single element array containing one hash
         # #YELLOW
-        slave_results.first.each do |key, value|
+        slave_results.fetch_hash.each_pair do |key, value|
           if metrics['general'].include?(key)
             # Replication lag being null is bad, very bad, so negativate it here
             value = -1 if key == 'Seconds_Behind_Master' && value.nil?
@@ -245,7 +239,7 @@ class Mysql2Graphite < Sensu::Plugin::Metric::CLI::Graphite
         variables_results = mysql.query('SHOW GLOBAL VARIABLES')
 
         category = 'configuration'
-        variables_results.each do |row|
+        variables_results.each_hash do |row|
           metrics[category].each do |metric, desc|
             if metric.casecmp(row['Variable_name']) == 0
               output "#{config[:scheme]}.#{mysql_shorthostname}.#{category}.#{desc}", row['Value']